Set up GHA workflow_dispatch for config generation (#11)
I'm not yet saying that this closes the related issue until it's fully
tested in `main` -- I'm not actually sure how to test this in a PR as
GHA doesn't seem to set the default inputs for the workflow unless it's
manually triggered in the `main` branch. (see:
https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/triggering-a-workflow#defining-inputs-for-manually-triggered-workflows)

But in theory, this should accept parameters from either a command line
workflow dispatch or the GUI that will get populated once this PR is
merged and create the job/task configs accordingly. I've also added a
schedule to run at 5am EST every Wednesday (but can change that easily
depending on project needs).