Use AzureCliCredential (#14)
When running the workflow in GHA, we're getting an authorization error
when uploading to blob storage. After debugging with @jkislin, we want
to check if the `DefaultAzureCredential` is logging in with the GHA
runner's SP instead of the CLI credentials that were provided in the
previous workflow step. This change should allow us to explicitly use
the NNH SP credentials and be able to push to storage.