Fix default workflow case (#17)
Closes #16 --

If no inputs are supplied to the GUI, then Github sets them on the
environment as empty string (idk why) -- before the workflow was failing
because `os.environ.get()` would get the empty string and set it for all
downstream variables. Now, if an empty string is set on the environment,
we use the default value instead, taking advantage of the `or` operator,
as empty strings evaluate to False: https://stackoverflow.com/a/1338532

Happy to do this another way if there are cleaner ways to handle!