Functionality to bulk update tasks (#36)
Closes #35 --

Adding a CLI command to recursively update task fields and apply them to
_all_ tasks within a given `job_id`. Also casting the list input for
`quantile_width` to float. Running this on a full set of configs takes a
bit of time (a couple of minutes), so I tested on a subset in this
container: `rt-epinow2-config /
Rt-estimation-2024-12-26T20-40-05.093193+00-00-9666be7ac3c911ef9673696322d524e5`

You can see that two sets of tasks were generated for each (state,
disease) pair with an updated timestamp in the `as_of_date` and task
filename. Feel free to stress test this and let me know if any issues
come up!

The command is `poetry run confmod bulk-update -j <job_id>`