Update default output container (#49)
## Goal
We want the config generator to send outputs to the `nssp-rt-testing`
container by default.

In a separate
[PR](https://github.com/CDCgov/cfa-epinow2-pipeline/pull/218), we have
edited the makefile in the cfa-epinow2-pipeline repo to set the output
container to `nssp-rt-v2` for production runs. We don't want the
production container to be the default, however, to avoid contaminating
production outputs with test or experimental outputs.

## Changes
- Replaced all instances of `zs-test-pipeline-update` with
`nssp-rt-testing`