running both workloads using runner-action (#53)
Based on the feedback from #52, I was able to rewrite the workloads to
run entirely via the runner-action which eliminates the containter app
job altogether and even skips the Github Artifact step from #50. I did
install `pyenv ` in the self-hosted runner image to make this work, but
I think that's a justifiable considering how many workflows are relying
on python.

Successfully ran run-workload.yaml:
https://github.com/CDCgov/cfa-config-generator/actions/runs/14091957948/job/39470557263

I did run the re-run-workload.yaml, but I don't know how to verify if it
worked or not:
https://github.com/CDCgov/cfa-config-generator/actions/runs/14092263286/job/39471586566

---------

Co-authored-by: Nate McIntosh <natemcintosh@gmail.com>